refactor(registry): extract pagination into a component
