+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
   "deno.enable": true,
   "deno.lint": true,
-  "deno.unstable": false
+  "deno.unstable": false,
+  "editor.defaultFormatter": "denoland.vscode-deno"
 }
